#159 - Filters = "" issue
I am going to recant my previous statement and will not be placing the filter value based on BCG values.

The reasoning behind this is there might be quite a few facilities where the BCG value is not supose to match the filter value. Also, in vERAM things that do not have a filter at all are not displayed (same thing happens in CRC). If I were to automatically place the bcg into the filter in the geojson, these facilities would have a bunch of things populate on screen that were never there before.

I personally don't know of a valid reason to set filters="" but I do know that some facilities do it.

Therefore, FEB will create the defaults in geojson with filters = [] and a message will be placed in the log so that users can quickly find what default filters dont have any filters in it. Then they can choose weather or not they want to add filters manually in the geojson. This result should be the same in CRC as it is in vERAM. (i.e. no filters = no display)